+++ authlib/oidc/core/grants/implicit.py
@@ -104,7 +104,7 @@ class OpenIDImplicitGrant(ImplicitGrant):
             if state:
                 params.append(("state", state))
         else:
-            error = AccessDeniedError(state=state)
+            error = AccessDeniedError()
             params = error.get_body()
 
         # http://openid.net/specs/oauth-v2-multiple-response-types-1_0.html#ResponseModes